requirements.txt
langchain==0.1.5
openai==1.12.0
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
chromadb==0.4.24
sentence-transformers==2.3.1
pydantic==2.6.1
mistralai==0.0.12
